% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildManual.R
\name{buildManual}
\alias{buildManual}
\title{Create a reference manual for an R package}
\usage{
buildManual(packageName, packagePath = "../")
}
\arguments{
\item{packageName}{The package name}

\item{packagePath}{The path where the package is located}
}
\description{
This function takes as input the name of the package and the path where its
folder is located. It assumes that the package folder contains a DESCRIPTION
file as well as a man folder with .Rd files generated using roxygen.
Currently, graphics and special fonts (i.e., Cyrillic) are not supported
}
\details{
This function uses \code{tools:::.DESCRIPTION_to_latex()} and \code{tools:::.Rdfiles2tex}, non-exported functions from \pkg{tools}.
}
